# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Praça.ai marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, best_sellers, rating, offers]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail with tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /checkout:
    post:
      operationId: processCheckout
      tags: [orders]
      summary: Process checkout and create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPct:
          type: ["integer", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        categorySlug:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        vendorLogoUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        salesCount:
          type: integer
        stock:
          type: integer
        isFavorited:
          type: boolean
        sizes:
          type: ["array", "null"]
          items:
            type: string
        deliveryDays:
          type: integer
        freeShipping:
          type: boolean
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - images
        - category
        - categorySlug
        - vendorId
        - vendorName
        - rating
        - reviewCount
        - salesCount
        - stock
        - isFavorited
        - deliveryDays
        - freeShipping
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            vendorRating:
              type: number
            vendorSalesCount:
              type: integer
            vendorDescription:
              type: ["string", "null"]
            shippingInfo:
              type: string
            returnPolicy:
              type: string
          required:
            - vendorRating
            - vendorSalesCount
            - shippingInfo
            - returnPolicy
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
      required:
        - products
        - total
        - page
        - hasMore
    Review:
      type: object
      properties:
        id:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: string
        date:
          type: string
        verified:
          type: boolean
        helpfulCount:
          type: integer
      required:
        - id
        - authorName
        - rating
        - comment
        - date
        - verified
        - helpfulCount
...
    CartItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
      required:
        - productId
        - quantity
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        productImageUrl:
          type: string
        quantity:
          type: integer
        price:
          type: number
        selectedSize:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - productImageUrl
        - quantity
        - price
    TrackingEvent:
      type: object
      properties:
        status:
          type: string
        description:
          type: string
        timestamp:
          type: string
        completed:
          type: boolean
      required:
        - status
        - description
        - timestamp
        - completed
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, shipped, out_for_delivery, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        deliveryAddress:
          type: string
        estimatedDelivery:
          type: string
        trackingCode:
          type: ["string", "null"]
        trackingEvents:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - items
        - subtotal
        - shipping
        - discount
        - total
        - paymentMethod
        - deliveryAddress
        - estimatedDelivery
        - trackingEvents
        - createdAt
    OrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        deliveryAddress:
          $ref: "#/components/schemas/AddressInput"
        deliveryOption:
          type: string
          enum: [express, standard, scheduled, pickup]
        paymentMethod:
          type: string
          enum: [credit_card, pix, boleto]
        couponCode:
          type: ["string", "null"]
      required:
        - items
        - deliveryAddress
        - deliveryOption
        - paymentMethod
    AddressInput:
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        complement:
          type: ["string", "null"]
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
      required:
        - street
        - number
        - neighborhood
        - city
        - state
        - zipCode
    CheckoutInput:
      type: object
      properties:
        deliveryAddress:
          $ref: "#/components/schemas/AddressInput"
        deliveryOption:
          type: string
          enum: [express, standard, scheduled, pickup]
        paymentMethod:
          type: string
          enum: [credit_card, pix, boleto]
        couponCode:
          type: ["string", "null"]
        cardHolder:
          type: ["string", "null"]
        cardNumber:
          type: ["string", "null"]
        cardExpiry:
          type: ["string", "null"]
        cardCvv:
          type: ["string", "null"]
      required:
        - deliveryAddress
        - deliveryOption
        - paymentMethod
    CheckoutResult:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        pixCode:
          type: ["string", "null"]
        pixQrCodeUrl:
          type: ["string", "null"]
        boletoUrl:
          type: ["string", "null"]
      required:
        - order
...